pyproject.toml
[project]
name = "curlviz"
version = "1.0.0"
description = "Visualization library for a curling sheet"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["skia-python>=87.6"]
